+++ pkg/gnet/sender.go
@@ -34,27 +34,27 @@ func SendToRole(msg proto.Message, sesID uint64, roleID uint64) {
 			zap.String("msgName", msgid.MsgIDS2C_name[int32(msgID)]),
 			zap.Any("data", msg),
 			zap.Any("to", pb.Server_Gateway),
-			zap.Int32("idx", serID),
+			zap.Uint8("idx", serID),
 			zap.Uint64("sessID", sesID),
 			zap.Uint64("roleID", roleID),
 		)
 	}
 }
 
-func SendToSrv(serType pb.Server, serID int32, msg proto.Message, roleID uint64, sesID uint64) {
+func SendToSrv(serType pb.Server, serID uint8, msg proto.Message, roleID uint64, sesID uint64) {
 	data, err := proto.Marshal(msg)
 	if err != nil {
-		zap.L().Warn("send msg error", zap.Error(err), zap.Any("serName", serType), zap.Int32("serID", serID))
+		zap.L().Warn("send msg error", zap.Error(err), zap.Any("serName", serType), zap.Uint8("serID", serID))
 		return
 	}
 	msgID, err := pb.GetMsgIDS2S(msg)
 	if err != nil {
-		zap.L().Warn("send msg error", zap.Error(err), zap.Any("serName", serType), zap.Int32("serID", serID))
+		zap.L().Warn("send msg error", zap.Error(err), zap.Any("serName", serType), zap.Uint8("serID", serID))
 		return
 	}
 	err = msgq.Q.Send(serType, serID, msgID, data, roleID, sesID)
 	if err != nil {
-		zap.L().Warn("send msg error", zap.Error(err), zap.Any("serName", serType), zap.Int32("serID", serID))
+		zap.L().Warn("send msg error", zap.Error(err), zap.Any("serName", serType), zap.Uint8("serID", serID))
 		return
 	}
 	if trace.Rule.ShouldLog(msgID, roleID, sesID) {
@@ -63,7 +63,7 @@ func SendToSrv(serType pb.Server, serID int32, msg proto.Message, roleID uint64,
 			zap.String("msgName", msgid.MsgIDS2S_name[int32(msgID)]),
 			zap.Any("data", msg),
 			zap.Any("to", serType),
-			zap.Int32("idx", serID),
+			zap.Uint8("idx", serID),
 			zap.Uint64("sessID", sesID),
 			zap.Uint64("roleID", roleID),
 		)
@@ -102,7 +102,7 @@ func SendToGate(msg proto.Message, sesID uint64) {
 	SendToSrv(pb.Server_Gateway, GateIDFromSesID(sesID), msg, 0, sesID)
 }
 
-func SendToGame(serID int32, msg proto.Message, sesID uint64, roleID uint64) {
+func SendToGame(serID uint8, msg proto.Message, sesID uint64, roleID uint64) {
 	SendToSrv(pb.Server_Game, serID, msg, roleID, sesID)
 }
