+++ pkg/gnet/servers.go
@@ -2,6 +2,6 @@ package gnet
 
 import "server/pkg/idgen"
 
-func GateIDFromSesID(gateID uint64) int32 {
-	return int32(idgen.MachineID(int64(gateID)))
+func GateIDFromSesID(gateID uint64) uint8 {
+	return uint8(idgen.MachineID(int64(gateID)))
 }
